Wrote all tests for IpValidator and moved the tests and the class into the correct namespace
